+++ src/main/java/com/whereyouad/WhereYouAd/domains/dashboard/presentation/DashboardController.java
@@ -4,14 +4,15 @@ import com.whereyouad.WhereYouAd.domains.dashboard.application.dto.response.Dash
 import com.whereyouad.WhereYouAd.domains.dashboard.domain.service.DashboardService;
 import com.whereyouad.WhereYouAd.domains.dashboard.presentation.docs.DashboardControllerDocs;
 import com.whereyouad.WhereYouAd.global.response.DataResponse;
+import jakarta.validation.Valid;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+
+import java.time.LocalDate;
 
 @RestController
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
@@ -30,4 +31,20 @@ public class DashboardController implements DashboardControllerDocs {
         return ResponseEntity.ok(DataResponse.from(budgetSummaryResponse));
     }
 
+    @GetMapping("/{orgId}/rankings/roas")
+    public ResponseEntity<DataResponse<DashboardResponse.RankingROASList>> getRoasRanking(
+            @AuthenticationPrincipal (expression = "userId") Long userId,
+            @PathVariable Long orgId,
+            @Valid @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate startDate,
+            @Valid @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate endDate)
+    {
+        // 조회 기간 설정이 없으면 오늘부터 최근 1개월 전 데이터 조회
+        LocalDate adjustedEnd = (endDate != null) ? endDate : LocalDate.now();
+        LocalDate adjustedStart = (startDate != null) ? startDate : adjustedEnd.minusMonths(1);
+
+        DashboardResponse.RankingROASList response = dashboardService.getRoasRanking(
+                userId, orgId, adjustedStart, adjustedEnd);
+
+        return ResponseEntity.ok(DataResponse.from(response));
+    }
 }
