+++ src/main/java/com/whereyouad/WhereYouAd/domains/dashboard/presentation/docs/DashboardControllerDocs.java
@@ -3,21 +3,56 @@ package com.whereyouad.WhereYouAd.domains.dashboard.presentation.docs;
 import com.whereyouad.WhereYouAd.domains.dashboard.application.dto.response.DashboardResponse;
 import com.whereyouad.WhereYouAd.global.response.DataResponse;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 
+import java.time.LocalDate;
+
 public interface DashboardControllerDocs {
 
-        @Operation(summary = "대시보드 - 예산 소진 현황 조회 API", description = "통합 대시보드는 요청 쿼리 파라미터 없이 호출, 플랫폼 대시보드는 요청 시 쿼리 파라미터에 providerType(GOOGLE, KAKAO, NAVER)과 함께 호출")
-        @ApiResponses({
-                        @ApiResponse(responseCode = "200", description = "성공"),
-                        @ApiResponse(responseCode = "401_1", description = "실패")
-        })
-        public ResponseEntity<DataResponse<DashboardResponse.BudgetSummaryResponse>> getBudgetSummary(
-                        @AuthenticationPrincipal(expression = "userId") Long userId,
-                        @RequestParam(name = "orgId") Long orgId,
-                        @RequestParam(required = false, name = "providerType") String providerType);
+    @Operation(
+            summary = "대시보드 - 예산 소진 현황 조회 API",
+            description = "통합 대시보드는 요청 쿼리 파라미터 없이 호출, 플랫폼 대시보드는 요청 시 쿼리 파라미터에 providerType(GOOGLE, KAKAO, NAVER)과 함께 호출"
+    )
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "401_1", description = "실패")
+    })
+    ResponseEntity<DataResponse<DashboardResponse.BudgetSummaryResponse>> getBudgetSummary(
+            @AuthenticationPrincipal(expression = "userId") Long userId,
+            @RequestParam(name = "orgId") Long orgId,
+            @RequestParam(required = false, name = "providerType") String providerType
+    );
+
+    @Operation(
+            summary = "ROAS 성과 순위 조회",
+            description = """
+            특정 조직(org)에 속한 모든 프로젝트의 광고 플랫폼(Provider)별 ROAS 성과 순위를 내림차순으로 반환합니다.
+
+            **[조회 기간(startDate, endDate) 설정]**
+            - 특정 기간을 지정하여 데이터를 조회할 수 있습니다. (예: 2026-03-01)
+            - 파라미터를 입력하지 않으면, 기본값으로 **최근 1개월(오늘 기준 한 달 전 ~ 오늘)**의 데이터를 조회합니다.
+
+            **[변화율(diffRate) 계산 기준]**
+            - 지정된 조회 기간과 **동일한 일수만큼의 직전 기간** 데이터를 바탕으로 ROAS 변화율(%)을 계산합니다.
+            - (예: 7일 치 데이터를 조회하면 직전 7일 치와 비교, 1개월 치 조회 시 직전 1개월 치와 비교)
+            - 이전 동일 기간에 비교할 기초 데이터(광고비 등)가 존재하지 않는 경우 `null`을 반환합니다.
+            """
+    )
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "성공 (결과 없으면 빈 배열)"),
+            @ApiResponse(responseCode = "400", description = "시작일이 종료일보다 늦은 경우"),
+            @ApiResponse(responseCode = "403", description = "해당 조직에 대한 접근 권한이 없는 경우"),
+            @ApiResponse(responseCode = "404", description = "존재하지 않는 조직")
+    })
+    ResponseEntity<DataResponse<DashboardResponse.RankingROASList>> getRoasRanking(
+            @AuthenticationPrincipal(expression = "userId") Long userId,
+            @Parameter(description = "조직 ID", required = true, example = "1") Long orgId,
+            @Parameter(description = "조회 시작일 (yyyy-MM-dd)") LocalDate startDate,
+            @Parameter(description = "조회 종료일 (yyyy-MM-dd)") LocalDate endDate
+    );
 }
